main pick and place package renamed to pick_and_place_application
