--- exercises/Perception-Driven_Manipulation/solution_ws/src/collision_avoidance_pick_and_place/launch/publish_sensor_data_bag.launch
@@ -1,7 +1,0 @@
-<?xml version="1.0"?>
-<launch>
-
-	<!--<param name="use_sim_time" value="True"/>-->
-	<node name="sensor_data" type="play" pkg="rosbag" args="$(find collision_avoidance_pick_and_place)/bag/sensor_data.bag -l "/> 
-
-</launch>
