--- exercises/Perception-Driven_Manipulation/solution_ws/src/collision_avoidance_pick_and_place/launch/ur5_generate_test_cloud_obstacles.launch
@@ -1,10 +1,0 @@
-<?xml version="1.0"?>
-<launch>
-
-	<!--<param name="use_sim_time" value="True"/>-->
-	<remap from="/generated_cloud" to="/kinect/depth_registered/points"/>
-	<node name="generate_cloud_node" pkg="collision_avoidance_pick_and_place" type="generate_cloud_node">
-		<rosparam command="load" file="$(find collision_avoidance_pick_and_place)/config/ur5/test_cloud_obstacles_description.yaml"/>
-	</node>
-
-</launch>
