--- exercises/Perception-Driven_Manipulation/solution_ws/src/collision_avoidance_pick_and_place/mainpage.dox
@@ -1,14 +1,0 @@
-/**
-\mainpage
-\htmlinclude manifest.html
-
-\b collision_avoidance_pick_and_place 
-
-<!-- 
-Provide an overview of your package.
--->
-
--->
-
-
-*/
