--- exercises/Perception-Driven_Manipulation/solution_ws/src/collision_avoidance_pick_and_place/src/tasks/create_motion_plan.cpp
@@ -1,60 +1,0 @@
-#include <collision_avoidance_pick_and_place/pick_and_place.h>
-
-/* CREATE MOTION PLAN
-  Goal:
-    - Creates a motion plan request using the desired end-effector pose and the current
-        robot state (joint configuration and payload status)
-    - Calls the moveit motion planning service and returns the motion plan if a valid one is
-        found.
-
-  Hints:
-
-*/
-
-namespace collision_avoidance_pick_and_place
-{
-
-bool PickAndPlaceApp::create_motion_plan(const geometry_msgs::msg::Pose &pose_target,
-    const moveit_msgs::msg::RobotState &start_robot_state_msg,
-    moveit_cpp::PlanningComponent::PlanSolution &plan_solution)
-{
-  // constructing motion plan goal constraints
-  std::vector<double> position_tolerances(3,0.01f);
-  std::vector<double> orientation_tolerances(3,0.01f);
-  geometry_msgs::msg::PoseStamped p;
-  p.header.frame_id = cfg.WORLD_FRAME_ID;
-  p.pose = pose_target;
-  moveit_msgs::msg::Constraints pose_goal = kinematic_constraints::constructGoalConstraints(
-      cfg.WRIST_LINK_NAME,
-      p,
-      position_tolerances,
-      orientation_tolerances);
-
-  // setting up planning configuration
-  moveit_cpp::PlanningComponent planning_component(cfg.ARM_GROUP_NAME, moveit_cpp);
-  moveit_cpp::PlanningComponent::PlanRequestParameters plan_parameters;
-  plan_parameters.load(node);
-  plan_parameters.planning_time = 60.0f;
-  plan_parameters.planning_attempts = 1;
-
-  // set planning goal
-  moveit::core::RobotState start_robot_state(moveit_cpp->getRobotModel());
-  moveit::core::robotStateMsgToRobotState(start_robot_state_msg, start_robot_state);
-  planning_component.setStartState(start_robot_state);
-  planning_component.setGoal({pose_goal});
-
-  // planning for goal
-  plan_solution = planning_component.plan(plan_parameters);
-  if(!plan_solution)
-  {
-    return false;
-  }
-
-  return true;
-}
-
-}
-
-
-
-
