--- exercises/Perception-Driven_Manipulation/solution_ws/src/collision_avoidance_pick_and_place/src/tasks/reset_world.cpp
@@ -1,23 +1,0 @@
-#include <collision_avoidance_pick_and_place/pick_and_place.h>
-
-/*    RESET WORLD
-  Goal:
-  Hints:
-*/
-void collision_avoidance_pick_and_place::PickAndPlaceApp::reset_world(bool refresh_octomap)
-{
-
-  // detach box if one is attached
-  moveit_msgs::msg::RobotState robot_state;
-  set_attached_object(false,geometry_msgs::msg::Pose(),robot_state);
-
-  // get new sensor snapshot
-  if(refresh_octomap)
-  {
-    detect_box_pick();
-  }
-
-  show_box(false);
-}
-
-
