+++ exercises/Perception-Driven_Manipulation/solution_ws/src/pick_and_place_application/config/ur5/pick_and_place_parameters.yaml
