+++ exercises/Perception-Driven_Manipulation/solution_ws/src/pick_and_place_application/config/ur5/target_recognition_parameters.yaml
