+++ exercises/Perception-Driven_Manipulation/solution_ws/src/pick_and_place_application/include/pick_and_place_application/pick_and_place_utilities.h
