+++ exercises/Perception-Driven_Manipulation/solution_ws/src/pick_and_place_application/launch/publish_sensor_data_bag.launch
@@ -1,0 +1,7 @@
+<?xml version="1.0"?>
+<launch>
+
+	<!--<param name="use_sim_time" value="True"/>-->
+	<node name="sensor_data" type="play" pkg="rosbag" args="$(find pick_and_place_application)/bag/sensor_data.bag -l "/> 
+
+</launch>
