+++ exercises/Perception-Driven_Manipulation/solution_ws/src/pick_and_place_application/launch/ur5_generate_test_cloud_obstacles.launch
@@ -1,0 +1,10 @@
+<?xml version="1.0"?>
+<launch>
+
+	<!--<param name="use_sim_time" value="True"/>-->
+	<remap from="/generated_cloud" to="/kinect/depth_registered/points"/>
+	<node name="generate_cloud_node" pkg="pick_and_place_application" type="generate_cloud_node">
+		<rosparam command="load" file="$(find pick_and_place_application)/config/ur5/test_cloud_obstacles_description.yaml"/>
+	</node>
+
+</launch>
