+++ exercises/Perception-Driven_Manipulation/solution_ws/src/pick_and_place_application/launch/ur5_pick_and_place.launch
@@ -1,0 +1,15 @@
+<?xml version="1.0"?>
+<launch>
+
+  <!-- pick and place node -->
+  <node pkg="pick_and_place_application" type="pick_and_place_node" name="pick_and_place_node" output="screen" required="true">
+    <rosparam command="load" file="$(find pick_and_place_application)/config/ur5/pick_and_place_parameters.yaml"/>
+  </node>
+
+  <include file="$(find ur5_collision_avoidance_moveit_config)/launch/move_group.launch">
+    <arg name="publish_monitored_planning_scene" value="true" />
+  </include>
+
+  <rosparam command="load" file="$(find ur5_collision_avoidance_moveit_config)/config/joint_names.yaml"/>
+
+</launch>
