+++ exercises/Perception-Driven_Manipulation/solution_ws/src/pick_and_place_application/launch/ur5_target_recognition.launch
@@ -1,0 +1,10 @@
+<?xml version="1.0"?>
+<launch>
+
+  <!-- target recognition node -->
+  <remap from="sensor_cloud" to="/kinect/depth_registered/points"/>
+  <node pkg="pick_and_place_application" type="target_recognition_service" name="target_recognition_service" output="screen">
+	<rosparam command="load" file="$(find pick_and_place_application)/config/ur5/target_recognition_parameters.yaml"/>
+  </node>
+
+</launch>
