+++ exercises/Perception-Driven_Manipulation/solution_ws/src/pick_and_place_application/mainpage.dox
@@ -1,0 +1,14 @@
+/**
+\mainpage
+\htmlinclude manifest.html
+
+\b pick_and_place_application 
+
+<!-- 
+Provide an overview of your package.
+-->
+
+-->
+
+
+*/
