+++ exercises/Perception-Driven_Manipulation/solution_ws/src/pick_and_place_application/src/nodes/generate_point_cloud.cpp
